wf-4-pushing-data-bq/code/requirements.txt
google-cloud-pubsub
google-cloud-storage
google-cloud-bigquery
functions-framework
xmltodict